etl/extract/requirements.txt
dlt[postgres]>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
